Patch 1.2
Separated DiscoverScreen from Home.js so that we can do things on DiscoverScreen better.
